+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/base/ui/BaseLoadingListFragment.java
@@ -18,6 +18,9 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.ninetwozero.bf4intel.R;
 import com.ninetwozero.bf4intel.factories.GsonProvider;
+import com.ninetwozero.bf4intel.ui.menu.RefreshEvent;
+import com.ninetwozero.bf4intel.utils.BusProvider;
+import com.squareup.otto.Subscribe;
 
 public abstract class BaseLoadingListFragment extends BaseListFragment implements Response.ErrorListener {
     protected Gson gson = GsonProvider.getInstance();
@@ -33,6 +36,19 @@ public abstract class BaseLoadingListFragment extends BaseListFragment implement
     public void onAttach(Activity activity) {
         super.onAttach(activity);
         requestQueue = Volley.newRequestQueue(activity);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        BusProvider.getInstance().register(this);
+        startLoadingData();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        BusProvider.getInstance().unregister(this);
     }
 
     @Override
@@ -53,23 +69,10 @@ public abstract class BaseLoadingListFragment extends BaseListFragment implement
         Log.w(getClass().getSimpleName(), "[onLoadFailure] " + error.getMessage());
     }
 
-    @Override
-    public void onCreateOptionsMenu(final Menu menu, final MenuInflater inflater) {
-        super.onCreateOptionsMenu(menu, inflater);
-
-        final MenuItem menuItem = menu.findItem(R.id.ab_action_refresh);
-        if ( menuItem != null ) {
-            menuItem.setVisible(true);
-        }
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(final MenuItem item) {
-        if (item.getItemId() == R.id.ab_action_refresh) {
-            startLoadingData();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
+    @Subscribe
+    public void onRefreshEvent(RefreshEvent event) {
+        Log.d("YOLO", getClass().getSimpleName() + " is refrehshing!");
+        startLoadingData();
     }
 
     protected <T> T fromJson(final String json, final Class<T> outClass) {
