+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/assignments/AssignmentGridFragment.java
@@ -13,6 +13,8 @@ import com.ninetwozero.bf4intel.json.assignments.Assignment;
 import com.ninetwozero.bf4intel.json.assignments.Assignments;
 import com.ninetwozero.bf4intel.network.SimpleGetRequest;
 import com.ninetwozero.bf4intel.resources.Keys;
+import com.ninetwozero.bf4intel.ui.menu.RefreshEvent;
+import com.squareup.otto.Subscribe;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -35,9 +37,8 @@ public class AssignmentGridFragment extends BaseLoadingFragment {
         return inflater.inflate(R.layout.fragment_assignments, container, false);
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
+    @Subscribe
+    public void onRefreshEvent(RefreshEvent event) {
         startLoadingData();
     }
