+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/awards/AwardGridFragment.java
@@ -15,6 +15,8 @@ import com.ninetwozero.bf4intel.json.awards.Medal;
 import com.ninetwozero.bf4intel.json.awards.Ribbon;
 import com.ninetwozero.bf4intel.network.SimpleGetRequest;
 import com.ninetwozero.bf4intel.resources.Keys;
+import com.ninetwozero.bf4intel.ui.menu.RefreshEvent;
+import com.squareup.otto.Subscribe;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -36,9 +38,8 @@ public class AwardGridFragment extends BaseLoadingFragment {
         return inflater.inflate(R.layout.fragment_assignments, container, false);
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
+    @Subscribe
+    public void onRefreshEvent(RefreshEvent event) {
         startLoadingData();
     }
