+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/menu/RefreshEvent.java
@@ -1,0 +1,5 @@
+package com.ninetwozero.bf4intel.ui.menu;
+
+public class RefreshEvent {
+    public RefreshEvent() {}
+}
